chore(extract): move styles/home into `<Home />` component
This commit is a continuation of placing `styled-components` that
are specific to a React component into the component source file
itself. This is for ease of refactoring if necessary and to simplify
the codebase.
