+++ public/serviceWorker.js
@@ -2,15 +2,10 @@ const CACHE_NAME = 'v1'
 const urlsToCache = [
   '.',
   '/assets/img/CreamLogo.png',
-  '/client/styles/GlobalStyles.js',
-  '/client/styles/home.js',
   '/client/styles/button.js',
-  '/client/styles/nav.js',
-  '/client/styles/footer.js',
-  '/client/styles/about.js',
-  '/client/styles/work.js',
-  '/client/styles/articles.js',
   '/client/styles/contact.js',
+  '/client/styles/GlobalStyles.js',
+  '/client/styles/work.js',
   'main.bundle.js',
   'index.html',
   'https://fonts.googleapis.com/css?family=Open+Sans:300,800',
